feat(GAT-8320): Refactors Federation god controller into service. Adds enabled_at flag to record first enable time of a federation
